+++ thebleep/shells/zsh.py
@@ -28,7 +28,7 @@ class Zsh(Generic):
                 TB_HISTORY="$(fc -ln -10)";
                 {fit_transport}
                 TB_CMD=$(
-                    TB_SHELL=zsh TB_ALIAS={name} TB_CAN_EDIT=1 TB_SHELL_ALIASES="$TB_SHELL_ALIASES" TB_HISTORY="$TB_HISTORY" thebleep {argument_placeholder} "$@"
+                    TB_SHELL=zsh TB_ALIAS={name} TB_CAN_EDIT=1 TB_SHELL_ALIASES="$TB_SHELL_ALIASES" TB_HISTORY="$TB_HISTORY" {command} {argument_placeholder} "$@"
                 );
                 TB_STATUS=$?;
                 if [ "$TB_STATUS" -eq {exit_edit} ]; then
@@ -41,6 +41,7 @@ class Zsh(Generic):
             }}
         '''.format(
             name=alias_name,
+            command=self._invocation(),
             argument_placeholder=ARGUMENT_PLACEHOLDER,
             fit_transport=fit_transport(),
             exit_edit=EXIT_EDIT,
@@ -88,9 +89,9 @@ class Zsh(Generic):
                 export THEBLEEP_INSTANT_MODE=True;
                 export THEBLEEP_OUTPUT_LOG={log};
                 trap 'rm -f {log}' EXIT HUP INT TERM;
-                thebleep --shell-logger {log};
+                {command} --shell-logger {log};
                 exit
-            '''.format(log=log_path)
+            '''.format(log=log_path, command=self._invocation())
 
     def _parse_alias(self, alias):
         name, value = alias.split('=', 1)